fix:修改用户表和实体类字段类型

+++ qbook-admin/src/main/java/cn/qinwh/qbookadmin/controller/AdminController.java
@@ -147,7 +147,7 @@ public class AdminController {
         User user = new User();
         user.setId(Integer.parseInt(bo.getId()));
         if(StringUtils.isNotEmpty(bo.getStatus())){
-            user.setStatus(Byte.parseByte(bo.getStatus()));
+            user.setStatus(Integer.parseInt(bo.getStatus()));
         }
         if(StringUtils.isNotEmpty(bo.getPassword())){
             user.setPassword(bo.getPassword());
